requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.49.0
jax==0.4.24
kiwisolver==1.4.5
matplotlib==3.8.3
mediapipe==0.10.10
ml-dtypes==0.3.2
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
RPi.GPIO==0.7.1
scipy==1.12.0
six==1.16.0
sounddevice==0.4.6
